requirements.txt
pandas>=2.1.4
psycopg2-binary>=2.9.9
openai>=2.7.0
python-dotenv>=1.0.1
tqdm>=4.66.1
numpy>=1.26.3
tiktoken>=0.12.0
httpx>=0.28.0
